refactor: rename _workspace_edit_mixin
